Moves Config::canonicalPath() to Sapi::canonicalPath()
