+++ Sources/Actions/Admin/Languages.php
@@ -28,6 +28,7 @@ use SMF\Lang;
 use SMF\Menu;
 use SMF\PackageManager\PackageUtils;
 use SMF\PackageManager\XmlArray;
+use SMF\Sapi;
 use SMF\SecurityToken;
 use SMF\Theme;
 use SMF\User;
@@ -905,7 +906,7 @@ class Languages implements ActionInterface
 		// Quickly load General language entries.
 		$old_txt = Lang::$txt;
 
-		require Config::canonicalPath($general_filename);
+		require Sapi::canonicalPath($general_filename);
 
 		Utils::$context['lang_file_not_writable_message'] = is_writable($general_filename) ? '' : Lang::getTxt('lang_file_not_writable', ['file' => $general_filename], file: 'ManageSettings');
