+++ Sources/Actions/Admin/Server.php
@@ -554,7 +554,7 @@ class Server implements ActionInterface
 			ACP::saveDBSettings($config_vars);
 
 			// Create the new directory, but revert to the previous one if anything goes wrong.
-			require_once Config::canonicalPath(Config::$sourcedir . '/Actions/Profile/Export.php');
+			require_once Sapi::canonicalPath(Config::$sourcedir . '/Actions/Profile/Export.php');
 			create_export_dir($prev_export_dir);
 
 			// Ensure we don't lose track of any existing export files.
