+++ Sources/Graphics/Gif/File.php
@@ -24,6 +24,7 @@ declare(strict_types=1);
 namespace SMF\Graphics\Gif;
 
 use SMF\Config;
+use SMF\Sapi;
 
 class File
 {
@@ -184,5 +185,5 @@ class File
 
 // 64-bit only functions?
 if (!\function_exists('smf_crc32')) {
-	require_once Config::canonicalPath(Config::$sourcedir . '/Subs-Compat.php');
+	require_once Sapi::canonicalPath(Config::$sourcedir . '/Subs-Compat.php');
 }
