+++ Sources/ItemList.php
@@ -332,7 +332,7 @@ class ItemList implements \ArrayAccess
 			$this->total_num_items = $this->options['get_count']['value'];
 		} else {
 			if (isset($this->options['get_count']['file'])) {
-				require_once Config::canonicalPath($this->options['get_count']['file']);
+				require_once Sapi::canonicalPath($this->options['get_count']['file']);
 			}
 
 			$call = Utils::getCallable($this->options['get_count']['function']);
@@ -483,7 +483,7 @@ class ItemList implements \ArrayAccess
 		} else {
 			// Get the file with the function for the item list.
 			if (isset($this->options['get_items']['file'])) {
-				require_once Config::canonicalPath($this->options['get_items']['file']);
+				require_once Sapi::canonicalPath($this->options['get_items']['file']);
 			}
 
 			$call = Utils::getCallable($this->options['get_items']['function']);
