+++ Sources/Theme.php
@@ -236,7 +236,7 @@ class Theme
 				$include = strtr(trim($include), ['$boarddir' => Config::$boarddir, '$sourcedir' => Config::$sourcedir, '$themedir' => $this->settings['theme_dir']]);
 
 				if (file_exists($include)) {
-					require_once Config::canonicalPath($include);
+					require_once Sapi::canonicalPath($include);
 				}
 			}
 		}
@@ -2347,9 +2347,9 @@ class Theme
 		$file_found = file_exists($filename);
 
 		if ($once && $file_found) {
-			require_once Config::canonicalPath($filename);
+			require_once Sapi::canonicalPath($filename);
 		} elseif ($file_found) {
-			require Config::canonicalPath($filename);
+			require Sapi::canonicalPath($filename);
 		}
 
 		if ($file_found !== true) {
