+++ Sources/Url.php
@@ -223,7 +223,7 @@ class Url implements \Stringable
 
 		if (!empty($this->host)) {
 			if (!\function_exists('idn_to_ascii')) {
-				require_once Config::canonicalPath(Config::$sourcedir . '/Subs-Compat.php');
+				require_once Sapi::canonicalPath(Config::$sourcedir . '/Subs-Compat.php');
 			}
 
 			// Convert the host using the Punycode algorithm
@@ -275,7 +275,7 @@ class Url implements \Stringable
 
 		if (!empty($this->host)) {
 			if (!\function_exists('idn_to_utf8')) {
-				require_once Config::canonicalPath(Config::$sourcedir . '/Subs-Compat.php');
+				require_once Sapi::canonicalPath(Config::$sourcedir . '/Subs-Compat.php');
 			}
 
 			// Decode the domain from Punycode.
@@ -713,7 +713,7 @@ class Url implements \Stringable
 
 			// Convert Punycode to Unicode
 			if (!\function_exists('idn_to_utf8')) {
-				require_once Config::canonicalPath(Config::$sourcedir . '/Subs-Compat.php');
+				require_once Sapi::canonicalPath(Config::$sourcedir . '/Subs-Compat.php');
 			}
 
 			foreach ($tlds as &$tld) {
